Refactor feff io: move code from __init__, split them into inputs and outputs modules. more cleanup and documentation
