+++ pymatgen/cli/feff_input_generation.py
@@ -3,13 +3,18 @@
 # Copyright (c) Pymatgen Development Team.
 # Distributed under the terms of the MIT License.
 
+from __future__ import division
 
 """
 Example script that generates FEFF input files from a cif file
 Remove comment # on write line to actually write files to disk
 """
 
-from __future__ import division
+import argparse
+
+from pymatgen.io.feff.sets import *
+from pymatgen.io.vasp import *
+from pymatgen.io.cif import CifParser
 
 __author__ = "Alan Dozier"
 __credits__= "Anubhav Jain, Shyue Ping Ong"
@@ -19,12 +24,6 @@ __maintainer__ = "Alan Dozier"
 __email__ = "adozier@uky.edu"
 __date__ = "April 7, 2013"
 
-import argparse
-
-from pymatgen.io.feff.sets import *
-from pymatgen.io.vasp import *
-from pymatgen.io.cif import CifParser
-
 
 def main():
     parser = argparse.ArgumentParser(description='''
@@ -33,10 +32,12 @@ def main():
     Version: 1.0
     Last updated: August, 2012''')
 
-
-    parser.add_argument('cif_file', metavar='cif_file', type=str, nargs=1, help='cif_file to use')
-    parser.add_argument('central_atom', metavar='central_atom', type=str, nargs=1, help='symbol of absorbing atom')
-    parser.add_argument('calc_type', metavar='calc_type', type=str, nargs=1, help='type of calc, currently XANES or EXAFS')
+    parser.add_argument('cif_file', metavar='cif_file', type=str, nargs=1,
+                        help='cif_file to use')
+    parser.add_argument('central_atom', metavar='central_atom', type=str,
+                        nargs=1, help='symbol of absorbing atom')
+    parser.add_argument('calc_type', metavar='calc_type', type=str, nargs=1,
+                        help='type of calc, currently XANES or EXAFS')
 
     args = parser.parse_args()
     cif_file = args.cif_file[0]
