+++ pymatgen/cli/feff_plot_cross_section.py
@@ -3,11 +3,10 @@
 # Copyright (c) Pymatgen Development Team.
 # Distributed under the terms of the MIT License.
 
-
 from __future__ import division
 
 """
-Script for plotting cross sections generated by FEFF found in xmu.dat files
+Script for plotting cross sections generated by FEFF found in xmu.dat files.
 """
 
 __author__ = "Alan Dozier"
@@ -20,12 +19,13 @@ __date__ = "April 7, 2013"
 
 import argparse
 
-from pymatgen.io.feff import Xmu
+from pymatgen.io.feff.outputs import Xmu
 from pymatgen.util.plotting_utils import get_publication_quality_plot
 
 
 def main():
-    parser = argparse.ArgumentParser(description='''Convenient DOS Plotter for Feff runs.
+    parser = argparse.ArgumentParser(description='''
+    Convenient DOS Plotter for Feff runs.
     Author: Alan Dozier
     Version: 1.0
     Last updated: April, 2013''')
