--- pymatgen/io/feff/tests/__init__.py
@@ -1,25 +1,0 @@
-# coding: utf-8
-# Copyright (c) Pymatgen Development Team.
-# Distributed under the terms of the MIT License.
-#!/usr/bin/env python
-
-from __future__ import division, unicode_literals
-
-"""
-#TODO: Write module doc.
-"""
-
-__author__ = 'Shyue Ping Ong'
-__copyright__ = 'Copyright 2013, The Materials Virtual Lab'
-__version__ = '0.1'
-__maintainer__ = 'Shyue Ping Ong'
-__email__ = 'ongsp@ucsd.edu'
-__date__ = '8/1/15'
-
-
-def main():
-    pass
-
-
-if __name__ == "__main__":
-    main()
